Integrates Stripe for payments
Connects to Stripe to handle payments between users, it:
- Upgrades Stripe library version.
- Introduces Stripe onboarding and payment processing.
- Creates customer sessions for managing payment methods.
- Sets up payment intents for secure transactions.
- Handles webhooks for account updates and payment confirmations.
- Removes card management entities as they are now managed by Stripe.

The overall flow is that a user (debtor) sets up payment methods with Stripe.
When a user creates a payment it generates a payment intent.
Once the payment intent succeds a webhook notifies the backend to create the payment in our db.